mikrobus: Add mikroBUS driver
Adds support for SPI mikroBUS addon boards with configuration based on
device tree. The goal is to get a minimal version in mainline to sort
out the device tree structure that should be used.

A mikroBUS board can use any combination of the following based protocols:
I2C, SPI, UART, PWM, Analog, GPIO with possibility of all pins being used
as GPIO instead of their original purpose. This requires the driver to be
flexible and identify the type of board based on the compatible string.

Provides a sysfs entry for each connector to enable passing the board
overlay for manual runtime discovery.

Signed-off-by: Ayush Singh <ayush@beagleboard.org>